+++ src/data/concepts/fix_concepts.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-# List of files to fix
-files="bonheur.ts connaissance.ts desespoir.ts desir.ts droit.ts egalite.ts realite.ts solidarite.ts sublime.ts verite.ts beaute.ts"
-
-for file in $files; do
-  echo "Processing $file..."
-  # This script will be run manually to fix each file
-done